refactor: change quantity and supplies to uint96 (#412)
**Motivation:**

Prepares the repo for https://github.com/llamaxyz/llama/issues/393#issuecomment-1607967564 to minimize the diff of the core changes

**Modifications:**

Change quantities and supplies from uint128 to uint96. Using uint96 lets us fit both total supply versions and a 64 bit timestamp into a single storage slot, which makes checkpointing total supply cheaper.

**Result:**

Diff on future PRs that implement the new checkpointing logic will be smaller.

+++ src/interfaces/ILlamaStrategy.sol
@@ -48,7 +48,7 @@ interface ILlamaStrategy {
   /// @param role The role to check quantity for.
   /// @param timestamp The timestamp at which to get the approval quantity.
   /// @return The quantity of the policyholder's approval.
-  function getApprovalQuantityAt(address policyholder, uint8 role, uint256 timestamp) external view returns (uint128);
+  function getApprovalQuantityAt(address policyholder, uint8 role, uint256 timestamp) external view returns (uint96);
 
   // -------- When Casting Disapproval --------
 
@@ -64,10 +64,7 @@ interface ILlamaStrategy {
   /// @param role The role to check quantity for.
   /// @param timestamp The timestamp at which to get the disapproval quantity.
   /// @return The quantity of the policyholder's disapproval.
-  function getDisapprovalQuantityAt(address policyholder, uint8 role, uint256 timestamp)
-    external
-    view
-    returns (uint128);
+  function getDisapprovalQuantityAt(address policyholder, uint8 role, uint256 timestamp) external view returns (uint96);
 
   // -------- When Queueing --------
